start unit tests, update timestamp creation, start flask work
Started writing unit tests - just one so far - to test the unix time
stamp creation function.  Determined that the tz handling may need to be
improved, and started that work - still need to fiddle with it to make
sure it will work as expected.

started to implement the flask app.  this is *barely* a start.  Followed
an online guide to setup a hello world app and will continue with that
until it's working with nginx locally.

When that's working, then try to get the app itself working via the web
server - perhaps a single function (e.g. add_user or get_holder).