continue crud tutorial: fixup db connection
working through the flask crud app tutorial I found on-line, specifically
fixed DB connection handling and defined a function to add a book to the
db (the example is a book title db).  Also added code to store the db
connection in the Flask 'g' object and to close the connection when
appropriate (still need to learn when that actually happens).