basic tos web app working
completed a flask crud app tutorial, then essentially converted it to
work as the tos app.  It's working, in the flask dev server, in a very
simplified way.  Specifically, when browsing the app you see: a list of
all events in the db, can add an event - with current timestamp, and
can delete an existing event.

Added code to handle add, delete and get events.  Reorged file layout.
Added templating (jinja2) for page rendering.  Added Flask code to handle
diplay of the main page (list events + add), and delete events.