fixed form processing on home page
the view function for root() previously called form.validate_on_submit() to determine if a form
was submitted in the request.  Apparently that function requires that validators be applied to
the flask_wtforms fields defined, which were not in my case - at least now at the moment.  By
switching to form.is_submitted() the form processing logic started working.  In case it's not clear
this is b/c the former case returned true only if the form was submitted AND VALIDATED.  The
latter returned true as long as formdata existed in the request.