fix systemd service configuration & some cosmetic changes
added the systemd service configuration for the app (tos.service).  This
was only on the production server, until now.  Ultimately, all app code
and config should be separate from the machine image.  Not sure yet where it
will all live.  Currently thining the repo is the place.  If so, will likely
need to add scripts to move things around on the server after cloning.

Added a readme file in the config directory - it mentions that it contains
various, possibly unrelated config files.

Also, renamed a couple of the nginx config files to be more meaningful (I hope)